EventQueue: refactor member naming and split of Event
