+++ src/core/IO/DiskWriterDriver.cpp
@@ -56,7 +56,7 @@ void* diskWriterDriver_thread( void* param )
 
 	DiskWriterDriver *pDriver = ( DiskWriterDriver* )param;
 
-	EventQueue::get_instance()->push_event( EVENT_PROGRESS, 0 );
+	EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, 0 );
 
 	auto pAudioEngine = Hydrogen::get_instance()->getAudioEngine();
 	
@@ -121,7 +121,7 @@ void* diskWriterDriver_thread( void* param )
 					.arg( pDriver->m_sFilename ).arg( sf_version_string() ) );
 		pDriver->m_bDoneWriting = true;
 		pDriver->m_bWritingFailed = true;
-		EventQueue::get_instance()->push_event( EVENT_PROGRESS, 100 );
+		EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, 100 );
 		pthread_exit( nullptr );
 		return nullptr;
 
@@ -163,7 +163,7 @@ void* diskWriterDriver_thread( void* param )
 					.arg( sf_version_string() ) );
 		pDriver->m_bDoneWriting = true;
 		pDriver->m_bWritingFailed = true;
-		EventQueue::get_instance()->push_event( EVENT_PROGRESS, 100 );
+		EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, 100 );
 		pthread_exit( nullptr );
 		return nullptr;
 	}
@@ -194,7 +194,7 @@ void* diskWriterDriver_thread( void* param )
 					.arg( sf_strerror( pSndfile ) ) );
 		pDriver->m_bDoneWriting = true;
 		pDriver->m_bWritingFailed = true;
-		EventQueue::get_instance()->push_event( EVENT_PROGRESS, 100 );
+		EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, 100 );
 		pthread_exit( nullptr );
 		return nullptr;
 	}
@@ -306,7 +306,7 @@ void* diskWriterDriver_thread( void* param )
 			// acquire the lock).
 			if ( ! pDriver->m_bIsRunning ) {
 				___ERRORLOG( "Driver was stop before export was completed." );
-				EventQueue::get_instance()->push_event( EVENT_PROGRESS, -1 );
+				EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, -1 );
 				pDriver->m_bWritingFailed = true;
 				tearDown();
 				return nullptr;
@@ -328,7 +328,7 @@ void* diskWriterDriver_thread( void* param )
 				if ( nMutexLockAttempts > 30 ) {
 					___ERRORLOG( "Too many attempts to lock the AudioEngine. Aborting." );
 					
-					EventQueue::get_instance()->push_event( EVENT_PROGRESS, -1 );
+					EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, -1 );
 					pDriver->m_bWritingFailed = true;
 					tearDown();
 					return nullptr;
@@ -397,7 +397,7 @@ void* diskWriterDriver_thread( void* param )
 							.arg( nBufferWriteLength )
 							.arg( sf_strerror( nullptr ) ) );
 
-				EventQueue::get_instance()->push_event( EVENT_PROGRESS, -1 );
+				EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, -1 );
 				pDriver->m_bWritingFailed = true;
 				tearDown();
 				return nullptr;
@@ -415,12 +415,12 @@ void* diskWriterDriver_thread( void* param )
 		int nPercent = static_cast<int>( ( float )(patternPosition +1) /
 										 ( float )nColumns * 100.0 );
 		if ( nPercent < 100 ) {
-			EventQueue::get_instance()->push_event( EVENT_PROGRESS, nPercent );
+			EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, nPercent );
 		}
 	}
 
 	// Explicitly mark export as finished.
-	EventQueue::get_instance()->push_event( EVENT_PROGRESS, 100 );
+	EventQueue::get_instance()->pushEvent( EVENT_PROGRESS, 100 );
 	
 	tearDown();
