+++ src/core/IO/JackAudioDriver.cpp
@@ -107,7 +107,7 @@ int JackAudioDriver::jackXRunCallback( void *arg ) {
 	// position.
 	JackAudioDriver::m_nIntegrationLastRelocationFrame = -1;
 #endif
-	EventQueue::get_instance()->push_event( EVENT_XRUN, 0 );
+	EventQueue::get_instance()->pushEvent( EVENT_XRUN, 0 );
 	return 0;
 }
 
@@ -550,7 +550,7 @@ void JackAudioDriver::relocateUsingBBT()
 		pAudioEngine->locate( fNewTick, false );
 	}
 
-	EventQueue::get_instance()->push_event( EVENT_RELOCATION, 0 );
+	EventQueue::get_instance()->pushEvent( EVENT_RELOCATION, 0 );
 
 	m_nTimebaseFrameOffset = pAudioEngine->getTransportPosition()->getFrame() -
 		m_JackTransportPos.frame;
@@ -655,7 +655,7 @@ void JackAudioDriver::updateTransportPosition()
 #endif
 
 				m_nTimebaseFrameOffset = 0;
-				EventQueue::get_instance()->push_event(
+				EventQueue::get_instance()->pushEvent(
 					EVENT_JACK_TIMEBASE_STATE_CHANGED,
 					static_cast<int>(m_timebaseState) );
 			}
@@ -674,7 +674,7 @@ void JackAudioDriver::updateTransportPosition()
 
 					m_timebaseState = Timebase::Listener;
 					m_nTimebaseFrameOffset = 0;
-					EventQueue::get_instance()->push_event(
+					EventQueue::get_instance()->pushEvent(
 						EVENT_JACK_TIMEBASE_STATE_CHANGED,
 						static_cast<int>(m_timebaseState) );
 				}
@@ -701,7 +701,7 @@ void JackAudioDriver::updateTransportPosition()
 
 					m_timebaseState = Timebase::None;
 					m_nTimebaseFrameOffset = 0;
-					EventQueue::get_instance()->push_event(
+					EventQueue::get_instance()->pushEvent(
 						EVENT_JACK_TIMEBASE_STATE_CHANGED,
 						static_cast<int>(m_timebaseState) );
 				}
@@ -1283,7 +1283,7 @@ void JackAudioDriver::initTimebaseControl()
 #endif
 
 			m_timebaseState = Timebase::Controller;
-			EventQueue::get_instance()->push_event(
+			EventQueue::get_instance()->pushEvent(
 				EVENT_JACK_TIMEBASE_STATE_CHANGED,
 				static_cast<int>(m_timebaseState) );
 		}
@@ -1327,7 +1327,7 @@ void JackAudioDriver::releaseTimebaseControl()
 				.arg( TimebaseToQString( m_timebaseState ) ) );
 #endif
 
-	EventQueue::get_instance()->push_event(
+	EventQueue::get_instance()->pushEvent(
 		EVENT_JACK_TIMEBASE_STATE_CHANGED, static_cast<int>(m_timebaseState) );
 
 }
@@ -1403,7 +1403,7 @@ void JackAudioDriver::JackTimebaseCallback(jack_transport_state_t state,
 
 		pDriver->m_timebaseState = Timebase::Controller;
 
-		EventQueue::get_instance()->push_event(
+		EventQueue::get_instance()->pushEvent(
 			EVENT_JACK_TIMEBASE_STATE_CHANGED,
 			static_cast<int>(pDriver->m_timebaseState) );
 	}
