+++ src/gui/src/Mixer/MasterLine.h
@@ -25,7 +25,6 @@
 #include <QtGui>
 #include <QtWidgets>
 
-#include <core/EventQueue.h>
 #include <core/Object.h>
 
 class Button;
