+++ src/gui/src/PatternEditor/PatternEditor.cpp
@@ -35,7 +35,6 @@
 #include <core/Basics/Song.h>
 #include <core/Hydrogen.h>
 #include <core/Preferences/Preferences.h>
-#include <core/EventQueue.h>
 #include <core/Basics/Drumkit.h>
 #include <core/Basics/Instrument.h>
 #include <core/Basics/InstrumentList.h>
