+++ src/gui/src/PatternEditor/PatternEditorSidebar.cpp
@@ -25,7 +25,6 @@
 
 #include <core/AudioEngine/AudioEngine.h>
 #include <core/CoreActionController.h>
-#include <core/EventQueue.h>
 #include <core/Hydrogen.h>
 #include <core/Basics/Drumkit.h>
 #include <core/Basics/Instrument.h>
