+++ src/gui/src/Widgets/MidiSenseWidget.cpp
@@ -131,7 +131,7 @@ void MidiSenseWidget::updateMidi(){
 					pAction );
 			}
 
-			H2Core::EventQueue::get_instance()->push_event( H2Core::EVENT_MIDI_MAP_CHANGED, 0 );
+			H2Core::EventQueue::get_instance()->pushEvent( H2Core::EVENT_MIDI_MAP_CHANGED, 0 );
 		}
 
 		close();
